feat: split callback methods into separate module
